Move attribute sources to the main linker namespace
